format ingredient quantities

+++ app/src/main/java/com/example/suzanne/recipesapp/models/IngredientSpecification.java
@@ -10,23 +10,23 @@ import android.os.Parcelable;
 public class IngredientSpecification implements Parcelable {
 
 
-    private int quantity;
+    private double quantity;
     private MeasurementType measure;
     private String ingredient;
 
-    public IngredientSpecification(int quantity, MeasurementType measure, String ingredient){
+    public IngredientSpecification(double quantity, MeasurementType measure, String ingredient){
         this.quantity = quantity;
         this.measure = measure;
         this.ingredient = ingredient;
     }
 
     public IngredientSpecification(Parcel parcel){
-        this.quantity = parcel.readInt();
+        this.quantity = parcel.readDouble();
         this.measure = MeasurementType.getType(parcel.readString());
         this.ingredient = parcel.readString();
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
@@ -45,7 +45,7 @@ public class IngredientSpecification implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel parcel, int i) {
-        parcel.writeInt(this.quantity);
+        parcel.writeDouble(this.quantity);
         parcel.writeString(this.measure.toString());
         parcel.writeString(this.ingredient);
     }
